tidied up control, removed /4, attack is now being set, tidied up wavetable, fixed performance issue around storing halfAmplitudes as float, switched up front calculations to double, switched note up front calculation to double

+++ src/note.h
@@ -12,7 +12,7 @@ public:
   uint16 tickCount;
 
   void init(float frequency) {
-    float fTicks = ticks_per_second / frequency;
+    double fTicks = ticks_per_second / frequency;
 
     ticks = round(fTicks);
     phasesPerTick = round(phases / fTicks);
